Add initial structure for domain error handler.
