+++ apps/server/src/core/error/domain/index.ts
@@ -1,0 +1,1 @@
+export { DomainErrorHandler } from './domainErrorHandler';
