+++ apps/server/src/core/index.ts
@@ -1,2 +1,3 @@
-export * from './core.module';
+export { CoreModule } from './core.module';
+export { DomainErrorHandler } from './error';
 export * from './interfaces';
