+++ apps/server/src/core/logger/index.ts
@@ -4,3 +4,4 @@ export * from './legacy-logger.service';
 export * from './logger';
 export * from './error-logger';
 export * from './types';
+export * from './logging.utils';
